Add rust unit tests for identity, reputation, document contracts
