+++ contracts/identity/src/lib.rs
@@ -104,85 +104,4 @@ impl IdentityContract {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use soroban_sdk::{
-        testutils::{Address as _, BytesN as _},
-        Env,
-    };
-
-    #[test]
-    fn test_register_and_verify() {
-        let env = Env::default();
-        env.mock_all_auths();
-
-        let contract_id = env.register(IdentityContract {}, ());
-        let client = IdentityContractClient::new(&env, &contract_id);
-
-        let admin = Address::generate(&env);
-        let wallet = Address::generate(&env);
-        let hash = BytesN::random(&env);
-
-        client.initialize(&admin);
-        client.register_identity(&hash, &wallet);
-
-        assert!(client.verify_identity(&wallet));
-        assert_eq!(client.get_user_identity(&wallet), hash);
-    }
-
-    #[test]
-    fn test_double_register_fails() {
-        let env = Env::default();
-        env.mock_all_auths();
-
-        let contract_id = env.register(IdentityContract {}, ());
-        let client = IdentityContractClient::new(&env, &contract_id);
-
-        let admin = Address::generate(&env);
-        let wallet = Address::generate(&env);
-        let hash = BytesN::random(&env);
-
-        client.initialize(&admin);
-        client.register_identity(&hash, &wallet);
-
-        let result = client.try_register_identity(&hash, &wallet);
-        assert_eq!(result, Err(Ok(IdentityError::AlreadyRegistered)));
-    }
-
-    #[test]
-    fn test_revoke_identity() {
-        let env = Env::default();
-        env.mock_all_auths();
-
-        let contract_id = env.register(IdentityContract {}, ());
-        let client = IdentityContractClient::new(&env, &contract_id);
-
-        let admin = Address::generate(&env);
-        let wallet = Address::generate(&env);
-        let hash = BytesN::random(&env);
-
-        client.initialize(&admin);
-        client.register_identity(&hash, &wallet);
-        assert!(client.verify_identity(&wallet));
-
-        client.revoke_identity(&wallet);
-        assert!(!client.verify_identity(&wallet));
-    }
-
-    #[test]
-    fn test_get_unregistered_fails() {
-        let env = Env::default();
-        env.mock_all_auths();
-
-        let contract_id = env.register(IdentityContract {}, ());
-        let client = IdentityContractClient::new(&env, &contract_id);
-
-        let admin = Address::generate(&env);
-        let wallet = Address::generate(&env);
-
-        client.initialize(&admin);
-
-        let result = client.try_get_user_identity(&wallet);
-        assert_eq!(result, Err(Ok(IdentityError::NotRegistered)));
-    }
-}
+mod test;
